+++ server/public/styles.css
@@ -1,0 +1,3 @@
+#logo {
+  color: black;
+}
